First clean compile
