+++ configure.ac
@@ -55,8 +55,8 @@ AM_INIT_AUTOMAKE([foreign subdir-objects])
 
 # When maintainer-mode is enabled, downstream users sometimes trip
 # trip over automake issues. See bug #649.
-AM_MAINTAINER_MODE([disable])
-# AM_MAINTAINER_MODE([enable])
+# AM_MAINTAINER_MODE([disable])
+AM_MAINTAINER_MODE([enable])
 
 AC_CONFIG_MACRO_DIR([m4])
