+++ link-grammar/dict-file/dictionary.c
@@ -22,6 +22,7 @@
 #include "dict-common/file-utils.h"
 #include "dict-common/idiom.h"
 #include "dict-common/regex-morph.h"
+#include "dict-ram/dict-ram.h"
 #include "post-process/pp_knowledge.h"
 #include "read-dialect.h"
 #include "read-dict.h"
