+++ link-grammar/dict-file/read-dict.h
@@ -15,22 +15,13 @@
 
 #include "dict-common/dict-structures.h"
 
-void print_dictionary_data(Dictionary dict);
-void print_dictionary_words(Dictionary dict);
-void print_dictionary_defines(Dictionary dict);
-
 Dictionary dictionary_six(const char *lang, const char *dict_name,
                           const char *pp_name, const char *cons_name,
                           const char *affix_name, const char *regex_name);
 Dictionary dictionary_create_from_file(const char *lang);
 bool read_dictionary(Dictionary dict);
 
-Dict_node * file_lookup_list(const Dictionary dict, const char *s);
-Dict_node * file_lookup_wild(Dictionary dict, const char *s);
-bool file_boolean_lookup(Dictionary dict, const char *s);
-void file_free_lookup(Dict_node *llist);
-
-void free_insert_list(Dict_node *ilist);
 void insert_list(Dictionary dict, Dict_node * p, int l);
+void free_insert_list(Dict_node *ilist);
 
 #endif /* _LG_READ_DICT_H_ */
